Add HTTP server functionality to main.go
The main.go script has been refactored to include an HTTP server function. The function is set to listen and serve on port 8080. The previously implemented functionality for adding, modifying, and iterating over a dynamic array has been commented out to make way for implementing the server. This is in line with shifting our app to a web-based interface, with data management moving from an array to a database or other complex storage system.